chatbot-service/requirements.txt
Flask==2.0.1
Werkzeug==2.3.8
google-generativeai==0.4.0
python-dotenv==0.21.0
